[#58] refactor: 파일 정리
UITextField에서만 사용되는 클래스와 구조체여서 extension: UITextField 위치로 변경했습니다.
